+++ CHANGELOG.md
@@ -88,6 +88,7 @@
 
 ### v3.7.5 / 2020-12-xx
 
+* [884](https://github.com/Vanessa219/vditor/issues/884) 工具栏自定义按钮禁用 `改进功能`
 * [877](https://github.com/Vanessa219/vditor/issues/877) 数学公式输入删除生成节点 `修复缺陷`
 * [882](https://github.com/Vanessa219/vditor/issues/882) 改进 HTML 转换 Markdown 时加粗、斜体等空格的处理 `改进功能`
 * [878](https://github.com/Vanessa219/vditor/issues/878) 移除列表标记符中文优化 `改进功能`
@@ -97,6 +98,7 @@
 * 文档修改
   * 3.7.5
     * `options.outline` 修改为 `{ enable: boolean, position: "left" | "right" }` 
+    * `toolbar.click` 参数修改为 `event: Event, vditor: IVditor`
   
 ### v3.7.4 / 2020-12-26